+++ ggc-core/src/ggc/SellTransaction.java
@@ -5,7 +5,7 @@ public class SellTransaction extends Transaction{
     private double _paymentPrice;
     private boolean _paid = false;
 
-    public SellTransaction(int id, int date, String productId, String partnerId, int quantity, float price, int paymentDeadline) {
+    public SellTransaction(int id, int date, String productId, String partnerId, int quantity, double price, int paymentDeadline) {
         super(id, date, productId, partnerId, quantity, price);
         _paymentDeadline = paymentDeadline;
         _paymentPrice = price;
